% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ASTM_E490_AM0.dt}
\alias{ASTM_E490_AM0.dt}
\title{Spectral data for the extraterrestrial solar spectrum ASTM E490}
\format{A data frame with  rows (119.5 nm to 1E6 nm, variable step) and 3 variables.}
\description{
Air Mass zero (AM0) spectrum
Dataset containing wavelengths from 119.5 nm to 1E6 nm and
tabulated values for spectral irradiance for the sun above the
terrestrial atmosphere ("mean" value).
}
\details{
The variables are as follows:
\itemize{
  \item w.length (nm)
  \item s.e.irrad (W m-2 nm-1)
  \item s.q.irrad (mol s-1 m-2 nm-1)
}
}
\note{
ASTM E490 (2000) AM0 standard (2004)
Extraterrestrial solar spectrum
Air mass zero (AM0)
\url{http://rredc.nrel.gov/solar/spectra/am0/ASTM2000.html}.
}
\examples{
library(photobiologySun)
library(photobiologyVIS)
with(ASTM_E490_AM0.dt, photon_ratio(w.length, s.e.irrad, Blue(), Green()))
plot(s.e.irrad~w.length, data=ASTM_E490_AM0.dt, main=comment(ASTM_E490_AM0.dt), type="l")
}
\keyword{datasets}

