% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{ASTM_G173_direct.data}
\alias{ASTM_G173_direct.data}
\alias{ASTM_G173_direct.data,}
\alias{ASTM_G173_global.data}
\title{Spectral data for the terrestrial solar spectrum ASTM G173}
\format{A data frame with  rows (119.5 nm to 1E6 nm, variable step) and 3 variables.}
\description{
Air Mass 1.5 (AM1.5) spectrum
Dataset containing wavelengths from 280.0 nm to 4000.0 nm and
tabulated values for spectral irradiance for the sun "typical for USA".
Dataframe ASTM_G173_direct.data contains the spectrum of direct solar
radiation on a plane at an inclination of 37 degrees, which together
with the assumed solar elevation results on a surface pointing at the
sun. Dataframe ASTM_G173_global.data conatains the spectral data for global
radiation on the same plane.
}
\details{
The variables are as follows:
\itemize{
  \item w.length (nm)
  \item s.e.irrad (W m-2 nm-1)
  \item s.q.irrad (mol s-1 m-2 nm-1)
}
}
\note{
ASTM G173-03 Reference Spectra Derived from SMARTS v. 2.9.2 AM1.5
Terrestrial solar spectrum
Air mass 1.5 (AM1.5)
(solar zenith angle 48.19 s)
from \url{http://rredc.nrel.gov/solar/spectra/am1.5/}
}
\examples{
library(photobiologySun)
library(photobiologyVIS)
with(sun_May_morning.data, photon_ratio(w.length, s.e.irrad, Blue(), Green()))
plot(s.e.irrad~w.length, data=sun_May_morning.data, main=comment(sun_May_morning.data), type="l")
}
\keyword{datasets}

