% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{Gueymard_AM0.data}
\alias{Gueymard_AM0.data}
\title{Spectral data for the extraterrestrial solar spectrum of Gueymard (2004)}
\format{A data frame with  rows (119.5 nm to 1E6 nm, variable step) and 3 variables.}
\description{
Air Mass zero (AM0) spectrum
Dataset containing wavelengths from 280.0 nm to 4000 nm and
tabulated values for spectral irradiance for the sun above the
terrestrial atmosphere ("mean" value). This spectrum was used
as the basis for calculating the terrestrial solar spectrum
defines by ASTM G173
}
\details{
The variables are as follows:
\itemize{
  \item w.length (nm)
  \item s.e.irrad (W m-2 nm-1)
  \item s.q.irrad (mol s-1 m-2 nm-1)
}
}
\note{
These data accompany the ASTM G173 spectral data.
Extraterrestrial solar spectrum
Air mass zero (AM0)
\url{http://rredc.nrel.gov/solar/spectra/am1.5/}.
}
\examples{
library(photobiologySun)
library(photobiologyVIS)
with(sun_May_morning.data, photon_ratio(w.length, s.e.irrad, Blue(), Green()))
plot(s.e.irrad~w.length, data=sun_May_morning.data, main=comment(sun_May_morning.data), type="l")
}
\keyword{datasets}

