% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.solar.irrad.data.r
\docType{data}
\name{ppfd_BF.data}
\alias{ppfd_BF.data}
\title{Ground level solar PAR photon irradiance, direct and diffuse (measured)}
\format{
A data frame with 24479 rows and 9 variables.
  variables.
}
\usage{
ppfd_BF.data
}
\description{
Dataset containing mean, maximum, minimum and standard deviation values for 
total, direct and difusse 
photosynthetically active radiation expresed as photon irradiance. Each value is a summary
60 consecutive readings acquired once every second.
}
\details{
The variables are as follows: 
\itemize{ 
\item time_EEST POSIXct Local time according to EET coordinates. 
\item ppfd_tot_mean numeric (umol m-2 m-2) 
\item ppfd_tot_min numeric (umol m-2 m-2) 
\item ppfd_tot_max numeric (umol m-2 m-2) 
\item ppfd_tot_sd numeric (umol m-2 m-2)
\item ppfd_diff_mean numeric (umol m-2 m-2) 
\item ppfd_diff_min numeric (umol m-2 m-2) 
\item ppfd_diff_max numeric (umol m-2 m-2) 
\item ppfd_diff_sd numeric (umol m-2 m-2)
}
}
\note{
Instrument used: Delta-T BF5 "quantum sensor" , mounted 
on tripod at approximately 2 m height. Data collected with a Campbell
Scientific CR6 datalogger, using analogue outputs from the sensor. 
Wavelength sensitivity range of the quantum sensor
is 400 nm to 700 nm. Location: Viikki campus of the University of Helsinki.
Coordinates: 60.226803 N, 25.019205 E.
}
\references{
\url{https://delta-t.co.uk/}
\url{https://www.campbellsci.com/}
}
\keyword{datasets}
