% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.solar.irrad.data.r
\docType{data}
\name{ppfd_LICOR.data}
\alias{ppfd_LICOR.data}
\title{Ground level solar PAR photon irradiance (measured)}
\format{
A data frame with 24479 rows and 5 variables.
}
\usage{
ppfd_LICOR.data
}
\description{
Dataset containing mean, maximum, minimum and standard deviation values for 
photosynthetically active radiaition expresed as photon irradiance. Each value is a summary
60 consecutive readings acquired once every second.
}
\details{
The variables are as follows: 
\itemize{ 
\item time_EEST POSIXct Local time according to EET coordinates. 
\item ppfd_mean numeric (umol m-2 m-2) 
\item ppfd_min numeric (umol m-2 m-2) 
\item ppfd_max numeric (umol m-2 m-2) 
\item ppfd_sd numeric (umol m-2 m-2)
}
}
\note{
Instrument used: LI-COR LI-190 quantum sensor,  mounted 
on tripod at approximately 2 m height. Data collected with a Campbell
Scientific CR6 datalogger. Sensor connected through a LI-COR millivolt adaptor
(604 ohm). Wavelength sensitivity range of the quantum sensor
is 400 nm to 700 nm. Location: Viikki campus of the University of Helsinki.
Coordinates: 60.226803 N, 25.019205 E.
}
\references{
\url{https://www.licor.com/env/}
\url{https://www.campbellsci.com/}
}
\keyword{datasets}
