% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{data}
\name{WMO_Wehrli_AM0.data}
\alias{WMO_Wehrli_AM0.data}
\title{Spectral data for the extraterrestrial solar spectrum of Wehrli}
\format{A data frame with  rows (119.5 nm to 1E6 nm, variable step) and 3 variables.}
\description{
Air Mass zero (AM0) spectrum
Dataset containing wavelengths from 199.5 nm to 10075 nm and
tabulated values for spectral irradiance for the sun above the
terrestrial atmosphere ("mean" value).
}
\details{
The variables are as follows:
\itemize{
  \item w.length (nm)
  \item s.e.irrad (W m-2 nm-1)
  \item s.q.irrad (mol s-1 m-2 nm-1)
}
}
\note{
Wehrli's (1985) AM0
extraterrestrial solar spectrum
Air Mass zero (AM0)
\url{http://rredc.nrel.gov/solar/spectra/am0/ASTM2000.html}.
}
\examples{
library(photobiologySun)
library(photobiologyVIS)
with(sun_May_morning.data, photon_ratio(w.length, s.e.irrad, Blue(), Green()))
plot(s.e.irrad~w.length, data=sun_May_morning.data, main=comment(sun_May_morning.data), type="l")
}
\keyword{datasets}

