% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.solar.irrad.data.r
\docType{data}
\name{irrad_Kipp.data}
\alias{irrad_Kipp.data}
\title{Ground level solar irradiance (measured)}
\format{A data frame with 24479 rows and 5 variables.
  variables.}
\usage{
irrad_Kipp.data
}
\description{
Dataset containing mean, maximum, minimum and standard deviation values for 
global radiation data expresed as (energy) irradiance. Each value is a summary
60 consecutive readings acquired once every second.
}
\details{
The variables are as follows: 
\itemize{ 
\item time_EEST POSIXct Local time according to EET coordinates. 
\item e_irrad_mean numeric (W m-2) 
\item e_irrad_min numeric (W m-2) 
\item e_irrad_max numeric (W m-2) 
\item e_irrad_sd numeric (W m-2)
}
}
\note{
Instrument used: Kipp SMP3 smart pyranometer, factory calibrated, mounted 
on tripod at approximately 2 m height. Data collected with a Campbell
Scientific CR6 datalogger. Wavelength sensitivity range of the pyranometer
is 300 nm to 2800 nm. Location: Viikki campus of the University of Helsinki.
Coordinates: 60.226803 N, 25.019205 E.
}
\references{
\url{https://www.kippzonen.com/}
}
\keyword{datasets}
