% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.solar.spectra.data.r
\docType{data}
\name{sun_hourly_august.spct}
\alias{sun_hourly_august.spct}
\title{Ground level spectral irradiance at hourly intervals}
\format{A data frame with 15841 rows (290 nm to 800 nm at 1 nm interval) and 3
  variables.}
\usage{
sun_hourly_august.spct
}
\description{
Datasets containing the wavelengths at 1.0 nm interval and tabulated values
spectral irradiance for the sun for 21 and 22 August in Helsinki, Finland.
}
\details{
The variables are as follows: 
\itemize{ 
\item w.length numeric (nm) 
\item s.e.irrad numeric (W m-2 nm-1) 
\item UTC POSIXct (UTC date and time)
}
The data set includes NAs for missing night-time spectral irardiance 
values.
}
\note{
The simualtion methods has been described in Lindfors et al. (2009).
}
\references{
Lindfors, A.; Heikkilä, A.; Kaurola, J.; Koskela, T. & Lakkala, K. (2009) 
Reconstruction of Solar Spectral Surface UV Irradiances Using Radiative 
Transfer Simulations. Photochemistry and Photobiology, 85: 1233–1239
}
\author{
Anders K. Lindfors (radiation transfer modelling)
}
\keyword{datasets}
